.Net core support (#63)
Bump to 1.6.0
Switched to VS2017 with new csproj format
Removed net3.5 support
Added netstandard1.3 and netstandard1.6 support
Global refactoring to support .Net Code API
Improved UT reliability

+++ MegaApiClient/Stream/MegaAesCtrStream.cs
@@ -201,7 +201,7 @@
         }
 
         // Copy to buffer
-        Array.Copy(output, 0, buffer, offset + pos - this.position, Math.Min(output.Length, this.streamLength - pos));
+        Array.Copy(output, 0, buffer, (int)(offset + pos - this.position), (int)Math.Min(output.Length, this.streamLength - pos));
 
         // Crypt to current chunk mac
         this.currentChunkMac = Crypto.EncryptAes(this.currentChunkMac, this.fileKey);
